WaterPvtThermal: move implementation to impl file
